pyproject.toml
[tool.poetry]
name = "async-firebase"
version = "5.1.0"
description = "Async Firebase Client - a Python asyncio client to interact with Firebase Cloud Messaging in an easy way."
license = "MIT"
authors = [
    "Oleksandr Omyshev <oomyshev@healthjoy.com>"
]
maintainers = [
    "Healthjoy Developers <developers@healthjoy.com>",
    "Oleksandr Omyshev <oomyshev@healthjoy.com>"
]
readme = "README.md"
repository = "https://github.com/healthjoy/async-firebase"
keywords = ["async", "asyncio", "firebase", "fcm", "python3", "push-notifications"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"

google-auth = "~2.38.0"
httpx = { version = ">=0.28.1,<1.0.0", extras = ["http2"] }
h11 = { version = ">=0.16.0" }

[tool.poetry.group.dev.dependencies]
pre-commit = "~4.2"
cryptography = "~44.0.2"
Faker = "~15.3"

# tests
pytest = "~8.3"
pytest-asyncio = "~0.26"
pytest-benchmark = "~5.1.0"
pytest-cov = "~6.1.1"
pytest-freezer = "~0.4.9"
pytest-httpx = "~0.35"
ruff = ">=0.9.6"
mypy = "~1.15.0"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["async_firebase", "tests"]
known-third-party = ["cryptography", "faker", "google", "httpx", "pytest", "pytest_httpx"]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "preserve"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
